+++ discovery_engine/lib/src/api/api.dart
@@ -53,6 +53,4 @@ export 'package:xayn_discovery_engine/src/domain/models/unique_id.dart'
 export 'package:xayn_discovery_engine/src/domain/models/view_mode.dart'
     show DocumentViewMode;
 export 'package:xayn_discovery_engine/src/domain/models/web_resource.dart'
-    show WebResource;
-export 'package:xayn_discovery_engine/src/domain/models/web_resource_provider.dart'
-    show WebResourceProvider;
+    show WebResource, WebResourceProvider;
