+++ discovery_engine/test/repository/hive_document_repo_test.dart
@@ -21,7 +21,7 @@ import 'package:xayn_discovery_engine/src/domain/models/document.dart'
 import 'package:xayn_discovery_engine/src/domain/models/unique_id.dart'
     show StackId;
 import 'package:xayn_discovery_engine/src/domain/models/web_resource.dart'
-    show WebResource;
+    show WebResource, WebResourceProvider;
 import 'package:xayn_discovery_engine/src/infrastructure/box_name.dart'
     show documentBox;
 import 'package:xayn_discovery_engine/src/infrastructure/repository/hive_document_repo.dart'
@@ -36,17 +36,22 @@ Future<void> main() async {
   final repo = HiveDocumentRepository();
 
   group('DocumentRepository', () {
-    final dummy = WebResource.fromJson(const <String, Object>{
-      'title': 'Example',
-      'displayUrl': 'domain.com',
-      'snippet': 'snippet',
-      'url': 'http://domain.com',
-      'datePublished': '1980-01-01T00:00:00.000000',
-      'provider': <String, String>{
-        'name': 'domain',
-        'thumbnail': 'http://thumbnail.domain.com',
-      },
-    });
+    final dummy = WebResource(
+      title: 'Example',
+      snippet: 'snippet',
+      url: Uri.parse('https://domain.com'),
+      displayUrl: Uri.parse('domain.com'),
+      datePublished: DateTime.utc(2022, 01, 01),
+      provider: WebResourceProvider(
+        name: 'domain',
+        thumbnail: Uri.parse('http://thumbnail.domain.com'),
+      ),
+      rank: 10,
+      score: 0.1,
+      country: 'en',
+      language: 'en',
+      topic: 'news',
+    );
     final stackId = StackId();
     final doc1 = Document(
       stackId: stackId,
